pyproject.toml
[project]
name = "chess-rag"
version = "0.1.0"
description = "Chess game analysis with RAG — AI commentary grounded in opening theory and pattern knowledge"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.35",
    "chess>=1.10",
    "faiss-cpu>=1.7,<1.8",
    "sentence-transformers>=2.2,<3.0",
    "groq>=0.9",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rag"]

[tool.uv]
constraint-dependencies = ["torch<=2.2.2", "numpy<2"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "requires_index: tests that need the FAISS index (run rag.ingestion first)",
]
